Change ulong to int

+++ Assets/LinearBeats/Scripts/Visuals/Scrolling/RailBehaviour.cs
@@ -7,12 +7,12 @@ namespace LinearBeats.Visuals
     {
         [SerializeField]
         private Rigidbody _rigidbody = null;
-        public ulong Pulse { get; set; } = 0;
+        public int Pulse { get; set; } = 0;
 
-        public void UpdateRailPosition(ulong currentPulse, float meterPerPulse)
+        public void UpdateRailPosition(int currentPulse, float meterPerPulse)
         {
-            //TODO: BPM 정지 구현 ulong a = _script.Timings[0].PulseStopDuration;
-            //TODO: BPM 역스크롤 구현 ulong a = _script.Timings[0].PulseReverseDuration (like a folded timeline!)
+            //TODO: BPM 정지 구현 int a = _script.Timings[0].PulseStopDuration;
+            //TODO: BPM 역스크롤 구현 int a = _script.Timings[0].PulseReverseDuration (like a folded timeline!)
             //TODO: 변속 대응하기
 
             float positionInMeter = meterPerPulse * (Pulse - currentPulse);
